refact: Relocating FillAndSmoothMixin
The bias_calc was getting too large so let's break it in parts to make
is easier to maintain and navigate. Creating a new sub-module mixins.
